python-controller/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
influxdb>=5.3.1
grpcio>=1.40.0
protobuf>=3.20.1
scikit-learn>=1.0
numpy>=1.21
pandas>=1.3